pyproject.toml
[tool.poetry]
name = "valve-diagnostic"
version = "0.1.0"
description = ""
authors = ["StigHaraldGustavsen <stighg@gmail.com>"]
readme = "README.md"
packages = [{include = "valve_diagnostic"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.0"
pandas = "^2.0.2"
matplotlib = "^3.7.1"
ipykernel = "^6.23.3"
cognite-sdk = "^6.4.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
